added logging moved compare_dicts to utils.py
